+++ lib/client/artifact_id_generator.go
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
- * All Rights Reserved.
- */
-
-package client