+++ lib/client/client.go
@@ -6,22 +6,18 @@
 package client
 
 import (
-	"crypto"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"github.com/kurtosis-tech/kurtosis-go/lib/client/artifact_id_provider"
 	"github.com/kurtosis-tech/kurtosis-go/lib/kurtosis_service"
 	"github.com/kurtosis-tech/kurtosis-go/lib/networks"
+	"github.com/kurtosis-tech/kurtosis-go/lib/services"
 	"github.com/kurtosis-tech/kurtosis-go/lib/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"os"
 	"strings"
 	"time"
-
-	// This is a special type of import that includes the correct hashing algorithm that we use
-	// If we don't have the "_" in front, Goland will complain it's unused
-	_ "golang.org/x/crypto/sha3"
 )
 
 const (
@@ -31,11 +27,19 @@ const (
 	// NOTE: right now this is hardcoded in the initializer as part of the contract between Kurtosis & a test suite image -
 	//  all test suite images MUST have this path available for mounting
 	suiteExecutionVolumeMountDirpath = "/suite-execution"
-
-	hashFunction = crypto.SHA3_256
 )
 
-func Run(testSuite testsuite.TestSuite, metadataFilepath string, servicesRelativeDirpath string, testName string, kurtosisApiIp string) int {
+type KurtosisClient struct {
+	artifactIdProvider artifact_id_provider.ArtifactIdProvider
+}
+
+func NewKurtosisClient() *KurtosisClient {
+	return &KurtosisClient{
+		artifactIdProvider: artifact_id_provider.NewDefaultArtifactIdProvider(),
+	}
+}
+
+func (client KurtosisClient) Run(testSuite testsuite.TestSuite, metadataFilepath string, servicesRelativeDirpath string, testName string, kurtosisApiIp string) int {
 	// Only one of these should be set; if both are set then it's an error
 	metadataFilepath = strings.TrimSpace(metadataFilepath)
 	testName = strings.TrimSpace(testName)
@@ -47,7 +51,7 @@ func Run(testSuite testsuite.TestSuite, metadataFilepath string, servicesRelativ
 	}
 
 	if !isMetadataFilepathEmpty {
-		if err := printSuiteMetadataToFile(testSuite, metadataFilepath); err != nil {
+		if err := client.printSuiteMetadataToFile(testSuite, metadataFilepath); err != nil {
 			logrus.Errorf("An error occurred writing test suite metadata to file '%v':", metadataFilepath)
 			fmt.Fprintln(logrus.StandardLogger().Out, err)
 			return errorExitCode
@@ -63,7 +67,7 @@ func Run(testSuite testsuite.TestSuite, metadataFilepath string, servicesRelativ
 			logrus.Error("Kurtosis API container IP argument was empty")
 			return errorExitCode
 		}
-		if err := runTest(servicesRelativeDirpath, testSuite, testName, kurtosisApiIp); err != nil {
+		if err := client.runTest(servicesRelativeDirpath, testSuite, testName, kurtosisApiIp); err != nil {
 			logrus.Errorf("An error occurred running test '%v':", testName)
 			fmt.Fprintln(logrus.StandardLogger().Out, err)
 			return errorExitCode
@@ -75,7 +79,7 @@ func Run(testSuite testsuite.TestSuite, metadataFilepath string, servicesRelativ
 // =========================================== Private helper functions ========================================
 // TODO Write tests for this by splitting it into metadata-generating function and writing function
 //  then testing the metadata-generating
-func printSuiteMetadataToFile(testSuite testsuite.TestSuite, filepath string) error {
+func (client KurtosisClient) printSuiteMetadataToFile(testSuite testsuite.TestSuite, filepath string) error {
 	logrus.Debugf("Printing test suite metadata to file '%v'...", filepath)
 	fp, err := os.OpenFile(filepath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
@@ -87,24 +91,25 @@ func printSuiteMetadataToFile(testSuite testsuite.TestSuite, filepath string) er
 	for testName, test := range testSuite.GetTests() {
 		testConfig := test.GetTestConfiguration()
 
-		// "Set" of used artifact URLs
+		// We create this "set" of used artifact URLs because the user could declare
+		//  multiple artifacts with the same URL
 		usedArtifactUrls := map[string]bool{}
 		for _, artifactUrl := range testConfig.FilesArtifactUrls {
 			usedArtifactUrls[artifactUrl] = true
 		}
 
-		artifactUrlsByHash := map[string]string{}
+		artifactUrlsById := map[string]string{}
 		for artifactUrl, _ := range usedArtifactUrls {
-			hexEncodedHash, err := hashArtifactUrl(artifactUrl)
+			artifactId, err := client.artifactIdProvider.GetArtifactId(artifactUrl)
 			if err != nil {
-				return stacktrace.Propagate(err, "An error occurred hashing artifact URL '%v'", artifactUrl)
+				return stacktrace.Propagate(err, "An error occurred getting the artifact ID for URL '%v'", artifactUrl)
 			}
-			artifactUrlsByHash[hexEncodedHash] = artifactUrl
+			artifactUrlsById[artifactId] = artifactUrl
 		}
 
 		testMetadata := NewTestMetadata(
 			testConfig.IsPartitioningEnabled,
-			artifactUrlsByHash)
+			artifactUrlsById)
 		allTestMetadata[testName] = *testMetadata
 	}
 	suiteMetadata := TestSuiteMetadata{
@@ -124,16 +129,6 @@ func printSuiteMetadataToFile(testSuite testsuite.TestSuite, filepath string) er
 	return nil
 }
 
-func hashArtifactUrl(artifactUrl string) (hexStr string, resultErr error) {
-	hasher := hashFunction.New()
-	artifactUrlBytes := []byte(artifactUrl)
-	if _, err := hasher.Write(artifactUrlBytes); err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred writing the artifact URL to the hash function")
-	}
-	hexEncodedHash := hex.EncodeToString(hasher.Sum(nil))
-	return hexEncodedHash, nil
-}
-
 /*
 Runs the single given test from the testsuite
 
@@ -147,7 +142,7 @@ Returns:
 	setupErr: Indicates an error setting up the test that prevented the test from running
 	testErr: Indicates an error in the test itself, indicating a test failure
 */
-func runTest(servicesRelativeDirpath string, testSuite testsuite.TestSuite, testName string, kurtosisApiIp string) error {
+func (client KurtosisClient) runTest(servicesRelativeDirpath string, testSuite testsuite.TestSuite, testName string, kurtosisApiIp string) error {
 	kurtosisService := kurtosis_service.NewDefaultKurtosisService(kurtosisApiIp)
 
 	tests := testSuite.GetTests()
@@ -163,7 +158,25 @@ func runTest(servicesRelativeDirpath string, testSuite testsuite.TestSuite, test
 		return stacktrace.Propagate(err, "An error occurred registering the test execution with the API container")
 	}
 
-	networkCtx := networks.NewNetworkContext(kurtosisService, suiteExecutionVolumeMountDirpath, servicesRelativeDirpath)
+	testConfig := test.GetTestConfiguration()
+	filesArtifactIdToGlobalArtifactId := map[services.FilesArtifactID]string{}
+	for filesArtifactId, url := range testConfig.FilesArtifactUrls {
+		globalArtifactId, err := client.artifactIdProvider.GetArtifactId(url)
+		if err != nil {
+			return stacktrace.Propagate(
+				err,
+				"An error occurred getting the global artifact ID for files artifact '%v' and url '%v'",
+				filesArtifactId,
+				url)
+		}
+		filesArtifactIdToGlobalArtifactId[filesArtifactId] = globalArtifactId
+	}
+
+	networkCtx := networks.NewNetworkContext(
+		kurtosisService,
+		suiteExecutionVolumeMountDirpath,
+		servicesRelativeDirpath,
+		filesArtifactIdToGlobalArtifactId)
 
 	logrus.Info("Setting up the test network...")
 	untypedNetwork, err := test.Setup(networkCtx)