+++ lib/testsuite/test_configuration.go
@@ -5,8 +5,7 @@
 
 package testsuite
 
-// The ID of a .tar.gz artifact containing files that will be used by one or more services
-type FilesArtifactID string
+import "github.com/kurtosis-tech/kurtosis-go/lib/services"
 
 /*
 Holds configuration values that, if set, give the test the ability to do special things
@@ -21,5 +20,5 @@ type TestConfiguration struct {
 
 	// A mapping of ID -> URL where the artifact containing files should be downloaded from
 	// The ID is the ID that service initializers will use when requesting to use the artifact
-	FilesArtifactUrls map[FilesArtifactID]string
+	FilesArtifactUrls map[services.FilesArtifactID]string
 }