+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 * Add tests for `RepartitionerBuilder` actions
 * Added extra testing inside `NetworkPartitionTest` to ensure that a node that gets added to a partition receives the correct blocking
 * Remove the HTTP client retrying from the JSON RPC client, because it can obscure errors like panics in the Kurtosis API and lead to red herring errors as it replays the call when the problem was the 
-* Add `FilesArtifactTest` to test pulling down a files artifact, mounting it inside a service, and using those files
+* Added the ability to mount external files into a service container:
+    * Added a new property, `FilesArtifactUrsl`, to `TestConfiguration` for defining files artifact URLs
+    * Add a new method, `GetFilesArtifactMountpoints`, to `DockerContainerInitializer` for defining where to mount external files artifacts
+    * Add `FilesArtifactTest` to test pulling down a files artifact, mounting it inside a service, and using those files
 
 ## 1.4.1
 * Point all old `kurtosis-docs` references to `docs.kurtosistech.com`